+++ tests/test_callback_for.py
@@ -1,7 +1,6 @@
 import scrapy
 import pytest
 from pytest_twisted import ensureDeferred
-from scrapy.utils.reqser import request_to_dict
 
 from web_poet.pages import ItemPage, ItemWebPage
 from scrapy_poet import (
@@ -89,7 +88,7 @@ def test_default_callback():
     """Sample request not specifying callback."""
     spider = MySpider()
     request = scrapy.Request('http://example.com/')
-    request_dict = request_to_dict(request, spider)
+    request_dict = request.to_dict(spider=spider)
     assert isinstance(request_dict, dict)
     assert request_dict['url'] == 'http://example.com/'
     assert request_dict['callback'] is None
@@ -99,13 +98,13 @@ def test_instance_method_callback():
     """Sample request specifying spider's instance method callback."""
     spider = MySpider()
     request = scrapy.Request('http://example.com/', callback=spider.parse_item)
-    request_dict = request_to_dict(request, spider)
+    request_dict = request.to_dict(spider=spider)
     assert isinstance(request_dict, dict)
     assert request_dict['url'] == 'http://example.com/'
     assert request_dict['callback'] == 'parse_item'
 
     request = scrapy.Request('http://example.com/', callback=spider.parse_web)
-    request_dict = request_to_dict(request, spider)
+    request_dict = request.to_dict(spider=spider)
     assert isinstance(request_dict, dict)
     assert request_dict['url'] == 'http://example.com/'
     assert request_dict['callback'] == 'parse_web'
@@ -117,7 +116,7 @@ def test_inline_callback():
     cb = callback_for(FakeItemPage)
     request = scrapy.Request('http://example.com/', callback=cb)
     with pytest.raises(ValueError) as exc:
-        request_to_dict(request, spider)
+        request.to_dict(spider=spider)
 
     msg = f'Function {cb} is not an instance method in: {spider}'
     assert str(exc.value) == msg
@@ -129,7 +128,7 @@ def test_inline_callback_async():
     cb = callback_for(FakeItemPageAsync)
     request = scrapy.Request('http://example.com/', callback=cb)
     with pytest.raises(ValueError) as exc:
-        request_to_dict(request, spider)
+        request.to_dict(spider=spider)
 
     msg = f'Function {cb} is not an instance method in: {spider}'
     assert str(exc.value) == msg