+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ TBR
       ``OverridesRegistry`` which provides a wide variety of features
       for better URL matching.
     * This resuls in a newer format in the ``SCRAPY_POET_OVERRIDES`` setting.
+* Removal of this deprecated module: ``scrapy.utils.reqser``
 
 * add ``async`` support for ``callback_for``.